Extract partnering family sort helpers
